requirements.txt
streamlit>=1.31.0,<1.37
pyarrow==14.0.2
python-dotenv>=1.0.1
openai>=1.40.0
requests>=2.31.0
beautifulsoup4>=4.12.3

